Small cleaning in Driver - intend to separate all CC(P) functions from standard ones. Weird variation in iteration count between identical EOMCC(P) runs doesn't seem to exist for CH+, maybe it was due to root flipping in H8.

+++ ccpy/utilities/utilities.py
@@ -20,7 +20,7 @@ def convert_excitations_c_to_f(excitations):
         if value.flags["F_CONTIGUOUS"]:
             continue
         else:
-            excitations[key] = np.asfortranarray(value, dtype=np.int32)
+            excitations[key] = np.asfortranarray(value)
     return excitations
 
 def print_memory_usage():
